+++ packages/app/cypress/e2e/top-nav.cy.ts
@@ -110,7 +110,7 @@ describe('App Top Nav Workflows', () => {
           expect(ctx.actions.browser.setActiveBrowserById).to.have.been.calledWith(browserId)
           expect(genId).to.eql('edge-chromium-stable')
           expect(ctx.actions.project.launchProject).to.have.been.calledWith(
-            ctx.coreData.currentTestingType, {}, undefined,
+            ctx.coreData.currentTestingType, undefined, undefined,
           )
         })
       })
@@ -547,7 +547,7 @@ describe('App Top Nav Workflows', () => {
             cy.contains('button', loginText.actionCancel).should('be.visible')
           })
 
-          cy.percySnapshot()
+          // cy.percySnapshot() // TODO: restore when Percy CSS is fixed. See https://github.com/cypress-io/cypress/issues/23435
 
           cy.withCtx((ctx) => {
             (ctx._apis.authApi.logIn as SinonStub).callsFake(async (onMessage) => {
@@ -594,7 +594,7 @@ describe('App Top Nav Workflows', () => {
             cy.contains('An unexpected error occurred').should('be.visible')
           })
 
-          cy.percySnapshot()
+          // cy.percySnapshot() // TODO: restore when Percy CSS is fixed. See https://github.com/cypress-io/cypress/issues/23435
 
           cy.findByRole('dialog', { name: loginText.titleFailed }).within(() => {
             cy.contains('button', loginText.actionTryAgain).should('be.visible')